removed folder classes, redundant moved logo into Hahnrich base class
