--- logo.txt
@@ -1,7 +1,0 @@
-#     #    #    #     # #     # ######  ###  #####  #     # 
-#     #   # #   #     # ##    # #     #  #  #     # #     # 
-#     #  #   #  #     # # #   # #     #  #  #       #     # 
-####### #     # ####### #  #  # ######   #  #       ####### 
-#     # ####### #     # #   # # #   #    #  #       #     # 
-#     # #     # #     # #    ## #    #   #  #     # #     # 
-#     # #     # #     # #     # #     # ###  #####  #     #
