--- src/Classes/Plugins/Twitch/config.json
@@ -1,6 +1,0 @@
-{
-    "clientId": "",
-    "clientSecret": "",
-    "faunaDomain": "db.eu.fauna.com",
-    "faunaToken": ""
-}
