+++ src/Plugin.ts
@@ -1,5 +1,5 @@
 import EventEmitter from "events"
-import Logger from "./Logger"
+import Logger from "./Utils/Logger"
 
 /**
  * The parent class of plugins to force them
