+++ src/Plugins/Discord/Commands/RandomQuoteCommand.ts
