+++ src/Plugins/Discord/Commands/RestartCommand.ts
