+++ src/Plugins/Discord/DiscordAddon.ts
@@ -1,5 +1,5 @@
 import EventEmitter from "events";
-import Logger from "../../Logger";
+import Logger from "../../Utils/Logger";
 import { Client } from "discord.js";
 
 export default abstract class DiscordAddon extends Logger {
