+++ src/Plugins/Discord/DiscordAddons.ts
